pyproject.toml
[tool.poetry]
name = "onyx-api-client"
version = "0.1.0"
description = "Onyx Rest and Websocket API client"
authors = ["lsbardel <lsbardella@onyxcapitaltechnology.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"
ruff = "^0.1.8"
isort = "^5.13.2"
mypy = "^1.7.1"
pytest-asyncio = "^0.23.5"
async-timeout = "^4.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = [
    "tests"
]

[tool.ruff]
select = ["E", "W", "F", "I", "B", "N"]
line-length = 88

[tool.mypy]
disallow_untyped_calls = false
disallow_untyped_defs = true
warn_no_return = true
